reminders timeouts set. Minimal refactor

+++ backend/src/resolvers/Mutation.js
@@ -1,15 +1,20 @@
 const { WebClient } = require('@slack/client');
 // An access token (from your Slack app or custom integration - xoxp, xoxb, or xoxa)
-const token = 'xoxb-18703011383-425949373157-pr8FbWDvwCX4LUojRBwmrhGV';
+require('dotenv').config({ path: 'variables.env' });
+const token = process.env.SLACK_TOKEN;
 const web = new WebClient(token);
 const moment = require('moment');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
-const { randomBytes } = require('crypto');
 const reminders = require('../slack/reminders');
+const getBotChannel = require('../utils/get-bot-channel');
 
 const Mutation = {
   async createPlayer(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     const player = await ctx.db.mutation.createPlayer(
       {
         data: { ...args }
@@ -20,6 +25,10 @@ const Mutation = {
     return player;
   },
   async createRinger(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     const username = args.name;
     const player = await ctx.db.mutation.createPlayer(
       {
@@ -31,16 +40,23 @@ const Mutation = {
     return player;
   },
   async deletePlayer(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     const player = await ctx.db.mutation.deletePlayer({
       where: { id: args.id }
     });
     return player;
   },
   async createMatch(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     // 1) Get time of day the reminder should be set
     const adminOptions = await ctx.db.query.adminOptions({ first: 1 });
-    const { reminderTime } = adminOptions[0];
-    console.log('\x1b[31m', 'admin options', adminOptions, '\x1b[0m');
+    const { reminderTime, reminderDay } = adminOptions[0];
 
     const reminderTimeArray = reminderTime.split(':');
     const hour = reminderTimeArray[0];
@@ -49,11 +65,21 @@ const Mutation = {
     // 2) Get time of match
     const matchDate = args.time;
 
-    // 3) Work out the reminder time based on admin set time of day
-    const reminder = moment(matchDate)
-      .subtract(1, 'day')
-      .hour(hour)
-      .minute(minute);
+    // 3) Work out the day of the week to set the reminder
+    let reminder;
+    const matchDayOfWeek = moment(matchDate).weekday();
+    if (reminderDay < matchDayOfWeek) {
+      reminder = moment(matchDate)
+        .weekday(reminderDay)
+        .hour(hour)
+        .minute(minute);
+    } else {
+      reminder = moment(matchDate)
+        .subtract(1, 'weeks')
+        .weekday(reminderDay)
+        .hour(hour)
+        .minute(minute);
+    }
 
     // 4) Store
     const match = await ctx.db.mutation.createMatch(
@@ -64,19 +90,29 @@ const Mutation = {
     );
 
     // 5) Trigger reminder to start in rest api that will send to slack
+    const Reminders = new reminders();
+    Reminders.setup(match);
 
+    // 6) Return match
     return match;
   },
   async deleteMatch(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     const player = await ctx.db.mutation.deleteMatch({
       where: { id: args.id }
     });
     return player;
   },
   async addToMatch(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     const where = { id: args.id };
     const playerId = args.playerId;
-
     return ctx.db.mutation.updateMatch(
       {
         data: {
@@ -91,7 +127,9 @@ const Mutation = {
   },
 
   async removeFromMatch(parent, args, ctx, info) {
-    console.log('\x1b[32m', 'args', args, '\x1b[0m');
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
 
     const where = { id: args.id };
     const playerId = args.playerId;
@@ -110,35 +148,32 @@ const Mutation = {
   },
 
   async saveSlackChannelMembers(parent, args, ctx, info) {
-    // First get all slack channels
-    const channelsInfo = await web.conversations.list();
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
 
-    // Lets get the chanel that the bot is added to
-    const channel = channelsInfo.channels.find(channel => {
-      return channel.is_member;
-    });
+    // 1) First get all slack channels
+    const botChannel = await getBotChannel().catch(e => e);
 
-    if (!channel)
-      throw new Error('There is no authenticated chanel for this slack app');
-
-    // Get the members list for that channel
+    // 3) Get the members list for that channel
     const channelInfo = await web.conversations.members({
-      channel: channel.id
+      channel: botChannel.id
     });
 
-    // Only way to get enough info about each member is to query all users
+    // 4) Only way to get enough info about each member is to query all users
     const allUsers = await web.users.list();
 
-    // Get only the users in the authenticated channel
+    // 5) Get only the users in the authenticated channel
     const membersInChannel = allUsers.members.filter(user =>
       channelInfo.members.includes(user.id)
     );
-    // Get rid of our bot user (or any others)
+    // 6) Get rid of our bot user (or any others)
     const nonBotUsers = membersInChannel.filter(user => !user.is_bot);
 
-    // Get current users in DB
+    // 7) Get current users in DB
     const savedPlayers = await ctx.db.query.players();
 
+    // 8) Create users that don't already exist
     const createNewUsers = async () => {
       const updatedIds = [];
       const promises = nonBotUsers.map(async user => {
@@ -175,6 +210,7 @@ const Mutation = {
       });
 
       await Promise.all(promises);
+
       return updatedIds;
     };
 
@@ -182,17 +218,23 @@ const Mutation = {
       throw new Error('error creating users', err);
     });
 
-    // Return a list of newly created users
+    // 10) Return a list of newly created users
     return await ctx.db.query.players({ where: { id_in: createdUsersIds } });
   },
   async setAdminOptions(parent, args, ctx, info) {
-    // We only ever set one adminOption
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
+    console.log('args', args);
+
+    // 1) We only ever set one adminOption so retrieve first
     const currentOptions = await ctx.db.query.adminOptions({ first: 1 });
     if (currentOptions.length === 0) {
       return ctx.db.mutation.createAdminOption({ data: { ...args } }, info);
     }
 
-    // If entry exists then we update
+    // 2) If entry exists then we update
     const existingOptionsId = currentOptions[0].id;
     return ctx.db.mutation.updateAdminOption(
       {
@@ -204,12 +246,12 @@ const Mutation = {
   },
 
   async signup(parent, args, ctx, info) {
-    // lowercase their email
+    // 1) lowercase their email
     args.email = args.email.toLowerCase();
-    // hash their password
+    // 2) hash their password
     const password = await bcrypt.hash(args.password, 10);
 
-    // create the user in the database
+    // 3) create the user in the database
     const user = await ctx.db.mutation.createUser(
       {
         data: {
@@ -220,14 +262,14 @@ const Mutation = {
       },
       info
     );
-    // create the JWT token for them
+    // 4) create the JWT token for them
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-    // We set the jwt as a cookie on the response
+    // 5) We set the jwt as a cookie on the response
     ctx.response.cookie('token', token, {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 365 // 1 year cookie
     });
-    // Finally we return the user to the browser
+    // 6) Finally we return the user to the browser
     return user;
   },
 
@@ -256,6 +298,10 @@ const Mutation = {
   },
 
   async sendReminder(parent, { matchId, playerId }, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     try {
       // 1) Get the match
       const match = await ctx.db.query.match({
@@ -271,11 +317,25 @@ const Mutation = {
 
       // 3) Lets trigger the reminder
       const Reminders = new reminders();
-      const response = await Reminders.sendMessage(slackId, match);
+      await Reminders.sendMessage(slackId, match);
       return 'Reminder sent';
     } catch (error) {
       return error;
     }
+  },
+
+  async updateUserType(parent, { id, type }, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
+    const player = await ctx.db.mutation.updatePlayer({
+      data: {
+        userType: type
+      },
+      where: { id }
+    });
+    return player;
   }
 };
